Moved all routine generation that relies on Tiled properties to CreepRoutineModel
