+++ SolStandard/SolStandard.csproj
@@ -415,7 +415,7 @@
     <Compile Include="Utility\Exceptions\UnitNotFoundInInitiativeListException.cs" />
     <Compile Include="Utility\Exceptions\VendorMisconfiguredException.cs" />
     <Compile Include="Utility\IResizable.cs" />
-    <Compile Include="Utility\Model\CreepModel.cs" />
+    <Compile Include="Utility\Model\CreepRoutineModel.cs" />
     <Compile Include="Utility\Monogame\BlankTexture.cs" />
     <Compile Include="Utility\Monogame\ISoundEffect.cs" />
     <Compile Include="Utility\Monogame\MusicBox.cs" />
